feat & refactor: Image 처리에 대한 예외를 따로 분리하여 ImageExceptionHandler에서 관리.
