+++ backend/src/main/java/com/ll/cafeservice/global/exception/GlobalExceptionHandler.java
@@ -2,6 +2,8 @@ package com.ll.cafeservice.global.exception;
 
 import com.ll.cafeservice.api.Empty;
 import com.ll.cafeservice.api.Result;
+import com.ll.cafeservice.global.exception.image.ImageStoreException;
+import com.ll.cafeservice.global.exception.image.ImageResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -16,17 +18,5 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(AccessDeniedException.class)
     public Result<Empty> handleAccessDeniedException(AccessDeniedException e) {
         return Result.error(403, "접근 권한이 없습니다.");
-    }
-
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(ResourceNotFoundException.class)
-    public Result<Empty> handleResourceNotFoundException(ResourceNotFoundException e) {
-        return Result.error(404, e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    @ExceptionHandler(ImageStoreException.class)
-    public Result<Empty> handleImageStoreException(ImageStoreException e) {
-        return Result.error(500, e.getMessage());
     }
 }
