+++ backend/src/main/java/com/ll/cafeservice/global/exception/image/ImageStoreException.java
@@ -1,4 +1,4 @@
-package com.ll.cafeservice.global.exception;
+package com.ll.cafeservice.global.exception.image;
 
 public class ImageStoreException extends RuntimeException {
     public ImageStoreException(String message) {
